internal: Initial bits of refactor to add online mode + use online mode for refresh
